# etc/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Homer-App User interface for WEB AI",
    "title": "Homer-App",
    "license": {
      "name": "AGPL",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "version": "1.1.2"
  },
  "host": "localhost:9080",
  "basePath": "/api/v3",
  "paths": {
...
    "/statistic/_db": {
      "get": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns data based upon filtered json",
        "produces": [
          "application/json"
        ],
        "tags": [
          "statistic"
        ],
        "operationId": "statisticGetStatisticDBList",
        "responses": {
          "200": {
            "description": "StatisticDb",
            "schema": {
              "$ref": "#/definitions/StatisticDb"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
    "/statistic/_measurements/{dbid}": {
      "get": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns data based upon filtered json",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "statistic"
        ],
        "operationId": "statisticGetStatisticMeasurementsList",
        "parameters": [
          {
            "type": "string",
            "description": "database id",
            "name": "dbid",
            "in": "path",
            "required": true,
            "schema": {
              "description": "database id",
              "type": "string"
            }
          },
          {
            "type": "object",
            "description": "StatisticSearchObject",
            "name": "StatisticSearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "StatisticSearchObject",
              "type": "object",
              "$ref": "#/definitions/StatisticSearchObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "StatisticObject",
            "schema": {
              "$ref": "#/definitions/StatisticObject"
            }
          }
        }
      }
    },
...
    "/statistic/_retentions": {
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns data based upon filtered json",
        "produces": [
          "application/json"
        ],
        "tags": [
          "statistic"
        ],
        "operationId": "statisticGetStatisticRetentionsList",
        "parameters": [
          {
            "type": "object",
            "description": "StatisticSearchObject",
            "name": "StatisticSearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "StatisticSearchObject",
              "type": "object",
              "$ref": "#/definitions/StatisticSearchObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "StatisticRetentions",
            "schema": {
              "$ref": "#/definitions/StatisticRetentions"
            }
          },
          "400": {
            "description": "UserLoginFailureResponse",
            "schema": {
              "$ref": "#/definitions/UserLoginFailureResponse"
            }
          }
        }
      }
    },
    "/statistic/_tags": {
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns data based upon filtered json",
        "produces": [
          "application/json"
        ],
        "tags": [
          "statistic"
        ],
        "operationId": "statisticGetStatisticTagsList",
        "parameters": [
          {
            "type": "object",
            "description": "StatisticSearchObject",
            "name": "StatisticSearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "StatisticSearchObject",
              "type": "object",
              "$ref": "#/definitions/StatisticSearchObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "StatisticDb",
            "schema": {
              "$ref": "#/definitions/StatisticDb"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
    "/statistic/data": {
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns data based upon filtered json",
        "produces": [
          "application/json"
        ],
        "tags": [
          "statistic"
        ],
        "operationId": "statisticStatisticData",
        "parameters": [
          {
            "type": "object",
            "description": "StatisticSearchObject",
            "name": "StatisticSearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "StatisticSearchObject",
              "type": "object",
              "$ref": "#/definitions/StatisticSearchObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "StatisticDb",
            "schema": {
              "$ref": "#/definitions/StatisticDb"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "FailureResponse": {
      "type": "object",
      "properties": {
        "error": {
          "description": "errot",
          "type": "string",
          "x-go-name": "Error"
        },
        "message": {
          "description": "message",
          "type": "string",
          "x-go-name": "Message"
        },
        "statuscode": {
          "description": "statuscode",
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusCode"
        }
      },
      "x-go-name": "UserTokenBadResponse",
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
...
    "StatisticDb": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "Results": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "Messages": {
                    "type": "object"
                  },
                  "Series": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "columns": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "x-go-name": "Columns",
                          "example": [
                            "name"
                          ]
                        },
                        "name": {
                          "type": "string",
                          "x-go-name": "Name",
                          "example": "databases"
                        },
                        "values": {
                          "type": "array",
                          "items": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "x-go-name": "Values",
                          "example": [
                            [
                              "telegraf"
                            ],
                            [
                              "_internal"
                            ],
                            [
                              "homer"
                            ]
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "x-go-name": "Data"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status",
          "example": "ok"
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total",
          "example": 1
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
    "StatisticObject": {
      "type": "object",
      "properties": {
        "param": {
          "type": "object",
          "properties": {
            "bfrom": {
              "type": "integer",
              "format": "int64",
              "x-go-name": "Bfrom"
            },
            "limit": {
              "type": "integer",
              "format": "int64",
              "x-go-name": "Limit"
            },
            "precision": {
              "type": "integer",
              "format": "int64",
              "x-go-name": "Precision"
            },
            "query": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "database": {
                    "type": "string",
                    "x-go-name": "Database"
                  },
                  "main": {
                    "type": "string",
                    "x-go-name": "Main"
                  },
                  "rawquery": {
                    "type": "string",
                    "x-go-name": "Rawquery"
                  },
                  "retention": {
                    "type": "string",
                    "x-go-name": "Retention"
                  },
                  "tag": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "x-go-name": "Tag"
                  },
                  "type": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "x-go-name": "Type"
                  }
                }
              },
              "x-go-name": "Query"
            },
            "total": {
              "type": "boolean",
              "x-go-name": "Total"
            }
          },
          "x-go-name": "Param"
        },
        "timestamp": {
          "type": "object",
          "properties": {
            "from": {
              "type": "integer",
              "format": "int64",
              "x-go-name": "From"
            },
            "to": {
              "type": "integer",
              "format": "int64",
              "x-go-name": "To"
            }
          },
          "x-go-name": "Timestamp"
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
    "StatisticRetentions": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "Results": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "Messages": {
                    "type": "object"
                  },
                  "Series": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "columns": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "x-go-name": "Columns",
                          "example": [
                            "name",
                            "duration",
                            "shardGroupDuration",
                            "replicaN",
                            "default"
                          ]
                        },
                        "values": {
                          "type": "array",
                          "items": {
                            "type": "array",
                            "items": {
                              "type": "object"
                            }
                          },
                          "x-go-name": "Values",
                          "example": [
                            [
                              "autogen",
                              "0s",
                              "168h0m0s",
                              1,
                              true
                            ]
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "x-go-name": "Data"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status",
          "example": "ok"
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total",
          "example": 1
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
    "StatisticSearchObject": {
      "type": "object",
      "properties": {
        "param": {
          "type": "object",
          "properties": {
            "bfrom": {
              "type": "integer",
              "format": "int64",
              "x-go-name": "Bfrom"
            },
            "limit": {
              "type": "integer",
              "format": "int64",
              "x-go-name": "Limit"
            },
            "precision": {
              "type": "integer",
              "format": "int64",
              "x-go-name": "Precision"
            },
            "search": {
              "type": "object",
              "properties": {
                "database": {
                  "type": "string",
                  "x-go-name": "Database"
                }
              },
              "x-go-name": "Search"
            },
            "total": {
              "type": "boolean",
              "x-go-name": "Total"
            }
          },
          "x-go-name": "Param"
        },
        "timestamp": {
          "type": "object",
          "properties": {
            "from": {
              "type": "integer",
              "format": "int64",
              "x-go-name": "From"
            },
            "to": {
              "type": "integer",
              "format": "int64",
              "x-go-name": "To"
            }
          },
          "x-go-name": "Timestamp"
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}